fixed issue in one of the new util functions

+++ Assets/Resources/Scripts/Tests/Utils/TestUtils.cs
@@ -32,7 +32,7 @@ public class TestUtils
 
     public static float GetCombinedFoodChangeInUpdateList(INonGenericCultureUpdate[] UpdateList)
     {
-        return UpdateList.Where(u => u.GetType() == typeof(FoodUpdate)).Sum(u => (int)u.GetCultureChange());
+        return UpdateList.Where(u => u.GetType() == typeof(FoodUpdate)).Sum(u => (float) u.GetCultureChange());
     }
 
     public static Color GetLastColorInUpdateList(INonGenericCultureUpdate[] UpdateList)
